[xtro] Fix processing common files when not all platforms are selected.
We can't process a common-fx.ignore file for a given framework if that
framework isn't included in any of the platforms we're building for.

Example: we can't process common-AppKit.ignore when only iOS is enabled,
because none of the errors listed in common-AppKit.ignore will be reported for
an iOS build.