[compare-commits] Remove argument to specify the GitHub comments file (#16310)
Remove the argument to specify the GitHub comments file, because it's overwritten with a computed value (i.e. completely ignored), and thus confusing what's supposed to happen in the calling script.

Also compute the default earlier, so that it's set to the expected value by the time we have error reporting that will try to write to the comments file.