[msbuild] Show an error when trying to build a binding project from Windows without a Mac connection.
Building a binding project from Windows without a Mac connection doesn't work,
because we need to execute bgen. Currently we just skip every task and target
that doesn't work from Windows, but that means nothing at all happens, which
is confusing (in particular if the binding project is referenced by other
projects, which will also fail to build).

So make it more explicit that a connection to a Mac is required to build a
binding project by showing an error when there's no connection.

Ref: https://github.com/xamarin/xamarin-macios/issues/16611