[msbuild] Fix libs reidentification deployed in a subdirectory
Dynamic libraries might be deployed in subdirectories such as libclrjit.dylib
from the nuget package cefglue.common:
Contents/MonoBundle/CefGlueBrowserProcess/libclrjit.dylib
The library ID for that library should be:
@executable_path/../MonoBundle/CefGlueBrowserProcess/libclrjit.dylib
Instead of:
@executable_path/../MonoBundle/libclrjit.dylib

Beside the library ID being wrong, when it's combined with the nuget
package microsoft.netcore.app.runtime.osx-x64 providing a library
with the same name, both uses the same `ReidentifiedPath`, which
can cause a failed in the InstallNameTool tasks that are run in parallel
operating on the same temporary file.

The following patch uses the `RelativePath` for the tempory file
used by `InstallNameTool` so that there are no clashes with other files
with the same name deployed in other directories. It also uses
the `RelativePath` to create the correct library id:
@executable_path/../../Contents/MonoBundle/CefGlueBrowserProcess/libclrjit.dylib