[DevOps] Checkout the yaml repos to fix an issue with the working dir.
Azure pipelines has this terrible design in which the path of the
checkout is different depending if you checkout a single repo or
several.

In this case, we have no issues on macios because we do know we have not
been checkout with anyother repo in the upload step, that is not the
case when working on the unified pipeline. Rather than adding some
complicated logic, we are going to be checking out the yaml templates so
that we have the same working directory structure.