[xtro] Fix targets file and print out full path to the html report.
There was a typo in the target name for creating the html report for .NET
('report-dotnet' vs 'dotnet-report').

Also print out the full path the html report when it's created, makes it much
easier to open the file from the command line because I can c&p the entire
path.