[generator] Unify generation/copying of Obsolete attributes.
Unify generation/copying of Obsolete attributes from api definitions to generated code.

This reduces code dupliation a bit, but it will also result in more
"[EditorBrowsable (Never)]" attributes (which we weren't generating in a few
cases).

In addition we're now copying Obsolete attributes for smart fields to the
generated code.