[tests] Add logic to detect when macOS and Mac Catalyst test apps don't launch properly. Fixes #xamarin/maccore@2414.
Implement a launch timeout for macOS and Mac Catalyst apps where if a certain environment
variable (LAUNCH_SENTINEL_FILE) is set, the app will create that file at launch.
The code launching the test app will wait 10 seconds and check if the file is there:
if it's not, something went wrong, in which case the app should be terminated and
launched again.

This necessitated re-implementing the launch script in C#, since it got quite complicated
to implement in bash.

This fixes an issue with Mac Catalyst apps where something would go wrong during
the app launch and nothing would happen (but the app wouldn't be deadlocked, it would
just sit there, doing nothing).

Fixes https://github.com/xamarin/maccore/issues/2414.