[maestro] Subscribe to the .NET 7 channel to get bumps for our own .NET 7 packages for .NET 8.
We need the latest released version numbers for the previous .NET version so
that we can reference and consume those when building projects using the
corresponding target framework.

Luckily we have a system for keeping dependencies up-to-date, so let's use it
to notify the current branch for any updates to the release branch for the
previous .NET version.