[msbuild] Add support for plugin library modules
Some projects like GStreamer, GIO or GdkPixbuf provide plugins as
dynamic libraries that are loaded on demand at runtime. These libraries
have usually the .dylib extension except for autotools-based projects
using libtool's -module flag that generate libraries with the .so extension.
These libraries need to be re-identified and deployed but that are not
linked to the application.