[generator] Fix property accessor attributes when one accessor's iOS availability is different than the property itself.
In the following scenario:

* A property is available on iOS (and implicitly Mac Catalyst).
* One of the property accessors is not available on iOS, but still available on Mac Catalyst (explicitly).

We'd generate the wrong availability attributes for ...

As for the fix: I can't really explain it. This code is rather impenetrable,
and the parameter names don't make much sense, but whatever I did seems to
work?

And it turns out this fix shows up in an existing test as well (the
generator's Bug35176 test), which I had to modify to remove the expectation of
(redundant) availability attributes that we no longer produce.