[CFNetwork] Add nullability to forgotten files.
Add nullability to the files. Although this files won't be present in
the new dotnet versions, it is good to add them for classic. Also, if we
move all files to have nullability enabled we should be able to pass it
as a flag and remove the need to use '#nullable enable' which people can
forget.