[release/7.0.2xx] [msbuild] Don't process assemblies over and over again in the UnpackLibraryResources target. Fixes #16377. (#17234)
Fixes #16377 

MSbuild has the ability to run `Target` partially. However in order to make sure we can use this, both the `Inputs` and `Outputs` need to have a 1-1 mapping. MSBuild will then use this mapping to figure out which files have actually changed. Then it will call the Target with only those files. 

In Xamarin.Android we make use of stamp files to provide this 1-1 mapping. We have a `pre` Target which will calculate an ItemGroup which will include a StampFile piece of metadata. We then use this new ItemGroup for both the `Inputs` and `Outputs`, thus meeting the 1-1 requirement. 

This PR updates the `_UnpackLibraryResources` to run partially. This will hopefully reduce the build time on incremental builds. 

Backport of #16416

Co-authored-by: Dean Ellis <dellis1972@googlemail.com>
Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>