[Generator] Add support for nullability. Part 1.
This is the first part of a 2 or more changes that will allow the APIs
to use ? as a way to mark a nullable field, parameter, property and
return method.

The way it works follows the documentation from the runtime that can be
found here: https://github.com/dotnet/roslyn/blob/main/docs/features/nullable-metadata.md

The tldr; of the documentation is as follows:

1. If we are looking at a value type, the compiler converts int? into
   Nullable<int>. This was already considered in the old code.
2. If we are inspecting a generic type we have to look for the
   information in two different attributes added by the compiler:
    - NullableAttribute: Each type reference in metadata may have an
        associated NullableAttribute with a byte[] where each byte
        represents nullability: 0 for oblivious, 1 for not annotated, and 2 for annotated.
    - NullableContextAttribute is valid in metadata on type and method declarations.
        The byte value represents the implicit NullableAttribute value for type reference
        within that scope that do not have an explicit NullableAttribute and would not
        otherwise be represented by an empty byte[].

The API is very similar to that provided  by dotnet6 but it allows us to
support classic. The move to the dotnet6 API should not be hard and
probably can be done by an IDE.

Once this API lands we will have to find the old usages of the check for
nullability and combine both. This new API should fix several issues we
have with nullability and nullability + generics.