[src] Unify build code for .NET platforms.
Use a make template for the build logic for our platform assemblies, so that we don't
unintentionally have slightly different build code for different platforms (especially
when making changes to the build this way ensures that we make the same change on all
platforms).

This also makes the code significantly shorter, and will make it easier to add more
platforms in the future if we ever need that.