[ObjCRuntime] Make BlockLiteral a ref struct in XAMCORE_5_0.
A ref struct has a few limitations that makes a lot of sense for blocks:
https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/ref-struct,
so by making BlockLiteral a ref struct, we're getting the C# compiler to help
us not writing broken code by accident.

The limitations also mean that it's a breaking change, so we can't do it
before XAMCORE_5_0.