[msbuild] Honor any existing values for HotRestartSignedAppOutputDir and HotRestartAppBundlePath.
This makes it easier to write tests that verify the files written in this paths,
because the tests can specify these output directories instead of evaluating a project
file or duplicating the logic to compute them.

Additionally, the tests can also easily make sure the output is cleaned up once the test is done.