[src] Use the new block API.
Update all block creation to use the new block API, where blocks are disposable.
This makes the code pattern a lot simpler.

I've changed all the P/Invokes to use an unsafe 'BlockLiteral*' pointer, because
'using' variables can't be passed as ref arguments, so the choice was either to make
the parameter type 'IntPtr' and cast away the pointer:

    using var block = new BlockLiteral ();
    PInvoke ((IntPtr) &block);

or make the parameter an unsafe 'BlockLiteral*' pointer:

    unsafe {
        using var block = new BlockLiteral ();
        PInvoke (&block);
    }

The upcoming support for function pointers don't have this choice: function pointers
are always unsafe, so I chose to go the unsafe route here as well, since it makes
the code simpler once support for function pointers has been implemented.