[dotnet] Include the parsed runtimeconfig.bin into ResolvedFileToPublish. Contributes towards #17579.
This makes it unnecessary to special-case this file for it to copied correctly
when building on Windows (once we've fixed the Windows build to use
ResolvedFileToPublish as the source of truth, like we do on macOS).

This is the first part of a fix for https://github.com/xamarin/xamarin-macios/issues/17579.