[AppKit] Make sure NSButtons are treated as toggled objects after setting a field. Fixes #17635.
If a wrapper type has a custom field (with a non-default value), we need to
mark the instance as dirty, to make it participate in the toggle-ref
machinery, and not get collected as long as the corresponding native instance
is around (otherwise the GC will collect the value in the field).

Fixes https://github.com/xamarin/xamarin-macios/issues/17635.