[msbuild] Set Hot Restart specific session id
In Hot Restart, we only use the session id to generate a unique temp path based on the project's current location.

By having a specific session id for Hot Restart, we disambiguate from the remote build session id that's used to know if a task needs to be executed remotely or locally.