[dotnet/msbuild] Copy binding resource packages to output directories together with CopyLocal assemblies. Fixes #13910.
With a project structure like this:

* Executable project references a library project.
* The library project references a binding project (or assembly).

The binding project's assembly will be copied to the library project's output directory
during the build. Unless we also make sure any binding resource packages are copied
as well, the executable project won't find those, and the final app won't contain
any native bits from the binding project.

The solution is to add any binding resource packages to the list of files to be copied
to the library's output directory.

Fixes https://github.com/xamarin/xamarin-macios/issues/13910.